pyproject.toml
[tool.poetry]
name = "tarkov-guide"
version = "0.1.0"
description = "A web application that helps players stacking quests in the online shooter Escape from Tarkov."
authors = ["Jordy van Vorselen <jordyvanvorselen@gmail.com>"]
license = "GNU Affero General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
pytest-django = "^4.5.2"
flake8-django = "^1.1.5"
django-jazzmin = "^2.5.0"
factory-boy = "^3.2.1"
djangorestframework = "^3.14.0"
pytest-clarity = "^1.0.1"

[tool.poetry.dev-dependencies]
mypy = "0.961"
django-stubs = {extras = ["compatible-mypy"], version = "^1.12.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
